implemented a kind of "login page" with local db implementation

+++ app/src/main/java/com/pillskeeper/data/Reminder.kt
@@ -2,5 +2,5 @@ package com.pillskeeper.data
 
 import java.util.*
 
-data class Reminder(var number_pills: Int, var minutes: Int, var hours: Int,
+data class Reminder(var number_pills: Float, var minutes: Int, var hours: Int,
                     var days: String, var duration: Date, var additionNotes: String?)
